requirements.txt
flask>=2.3,<3
numpy>=1.26
pandas>=2.1
scikit-learn>=1.3
# Make TensorFlow optional; if installed, the app will use deep model
# tensorflow>=2.16; python_version < '3.12'
pillow>=10.0
matplotlib>=3.8
opencv-python>=4.9
Werkzeug>=2.3
itsdangerous>=2.1
Jinja2>=3.1
MarkupSafe>=2.1
click>=8.1